+++ addon/models/model.js
@@ -1,16 +1,14 @@
 import Ember from 'ember';
 import DS from 'ember-data';
 import createProj from '../utils/create';
-import EmberValidations from 'ember-validations';
 
 /**
-  Base model that supports projections and validations.
+  Base model that supports projections.
 
   @module ember-flexberry-data
   @class Model
   @namespace Projection
   @extends DS.Model
-  @uses EmberValidationsMixin
   @uses Ember.EventedMixin
 
   @event preSave
@@ -18,8 +16,8 @@ import EmberValidations from 'ember-validations';
   @param {Promise[]} promises Array to which custom 'preSave' promises could be pushed
 
   @public
- */
-var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
+*/
+let Model = DS.Model.extend(Ember.Evented, {
   /**
     Stored canonical `belongsTo` relationships.
 
@@ -28,15 +26,6 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
     @private
   */
   _canonicalBelongsTo: Ember.computed(() => ({})),
-
-  /**
-    Model validation rules.
-
-    @property validations
-    @type Object
-    @default {}
-  */
-  validations: {},
 
   /**
     Flag that indicates sync up process of model is processing.
@@ -75,56 +64,6 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
   isDestroyedDuringSyncUp: false,
 
   /**
-    Checks that model satisfies validation rules defined in 'validations' property.
-
-    @method validate
-    @param {Object} [options] Method options
-    @param {Boolean} [options.validateDeleted = true] Flag: indicates whether to validate model, if it is deleted, or not
-    @return {Promise} A promise that will be resolved if model satisfies validation rules defined in 'validations' property
-  */
-  validate(options) {
-    options = Ember.merge({ validateDeleted: true }, options || {});
-    if (options.validateDeleted === false && this.get('isDeleted')) {
-      // Return resolved promise, because validation is unnecessary for deleted model.
-      return new Ember.RSVP.Promise((resolve) => {
-        resolve();
-      });
-    }
-
-    // Validate model.
-    let validationPromises = {
-      base: this._super(options)
-    };
-
-    let hasManyRelationships = Ember.A();
-    this.eachRelationship((name, attrs) => {
-      if (attrs.kind === 'hasMany') {
-        hasManyRelationships.pushObject(attrs.key);
-      }
-    });
-
-    // Validate hasMany relationships.
-    hasManyRelationships.forEach((relationshipName) => {
-      let details = this.get(relationshipName);
-      if (!Ember.isArray(details)) {
-        details = Ember.A();
-      }
-
-      details.forEach((detailModel, i) => {
-        validationPromises[relationshipName + '.' + i] = detailModel.validate.apply(detailModel, options);
-      });
-    });
-
-    return new Ember.RSVP.Promise((resolve, reject) => {
-      Ember.RSVP.hash(validationPromises).then((hash) => {
-        resolve(this.get('errors'));
-      }).catch((reason) => {
-        reject(this.get('errors'));
-      });
-    });
-  },
-
-  /**
     Triggers model's 'preSave' event & allows to execute some additional async logic before model will be saved.
 
     @method beforeSave
@@ -158,7 +97,7 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
   },
 
   /**
-    Validates model, triggers 'preSave' event, and finally saves model.
+    The 'preSave' event is called and finally saves model.
 
     @method save
 
@@ -171,11 +110,7 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
     options = Ember.merge({ softSave: false }, options || {});
 
     return new Ember.RSVP.Promise((resolve, reject) => {
-      // If we are updating while syncing up then checking of validation rules should be skipped
-      // because they can be violated by unfilled fields of model.
-      let promise = this.get('isSyncingUp') && this.get('dirtyType') === 'updated' ?
-        Ember.RSVP.resolve() : this.validate({ validateDeleted: false });
-      promise.then(() => this.beforeSave(options)).then(() => {
+      this.beforeSave(options).then(() => {
         // Call to base class 'save' method with right context.
         // The problem is that call to current save method will be already finished,
         // and traditional _this._super will point to something else, but not to DS.Model 'save' method,
@@ -187,13 +122,12 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
         // Assuming that record is not updated during sync up;
         this.set('isUpdatedDuringSyncUp', false);
 
-        // Model validation was successful (model is valid or deleted),
-        // all 'preSave' event promises has been successfully resolved,
+        // All 'preSave' event promises has been successfully resolved,
         // finally model has been successfully saved,
         // so we can resolve 'save' promise.
         resolve(value);
       }).catch(reason => {
-        // Any of 'validate', 'beforeSave' or 'save' promises has been rejected,
+        // Any of 'beforeSave' or 'save' promises has been rejected,
         // so we should reject 'save' promise.
         reject(reason);
       });
@@ -353,24 +287,6 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
   },
 
   /**
-    Initializes model.
-  */
-  init() {
-    this._super(...arguments);
-
-    // Attach validation observers for hasMany relationships.
-    this.eachRelationship((name, attrs) => {
-      if (attrs.kind !== 'hasMany') {
-        return;
-      }
-
-      let detailsName = attrs.key;
-      Ember.addObserver(this, `${detailsName}.[]`, this, this._onChangeHasManyRelationship);
-      Ember.addObserver(this, `${detailsName}.@each.isDeleted`, this, this._onChangeHasManyRelationship);
-    });
-  },
-
-  /**
     Fired when the record is loaded from the server.
     [More info](http://emberjs.com/api/data/classes/DS.Model.html#event_didLoad).
 
@@ -401,78 +317,6 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
   didCreate() {
     this._super(...arguments);
     this._saveCanonicalBelongsTo();
-  },
-
-  /**
-    Destroys model.
-  */
-  willDestroy() {
-    this._super(...arguments);
-
-    // Attach validation observers for hasMany relationships.
-    this.eachRelationship((name, attrs) => {
-      if (attrs.kind !== 'hasMany') {
-        return;
-      }
-
-      let detailsName = attrs.key;
-      Ember.removeObserver(this, `${detailsName}.[]`, this, this._onChangeHasManyRelationship);
-      Ember.removeObserver(this, `${detailsName}.@each.isDeleted`, this, this._onChangeHasManyRelationship);
-    });
-  },
-
-  /**
-    Observes & handles changes in each hasMany relationship.
-
-    @method _onChangeHasManyRelationship
-    @param {Object} changedObject Reference to changed object.
-    @param {changedPropertyPath} changedPropertyPath Path to changed property.
-    @private
-  */
-  _onChangeHasManyRelationship(changedObject, changedPropertyPath) {
-    Ember.run.once(this, '_aggregateHasManyRelationshipValidationErrors', changedObject, changedPropertyPath);
-  },
-
-  /**
-    Aggregates validation error messages for hasMany relationships.
-
-    @method _aggregateHasManyRelationshipValidationErrors
-    @param {Object} changedObject Reference to changed object.
-    @param {changedPropertyPath} changedPropertyPath Path to changed property.
-    @private
-  */
-  _aggregateHasManyRelationshipValidationErrors(changedObject, changedPropertyPath) {
-    // Retrieve aggregator's validation errors object.
-    let errors = Ember.get(this, 'errors');
-
-    let detailsName = changedPropertyPath.split('.')[0];
-    let details = Ember.get(this, detailsName);
-    if (!Ember.isArray(details)) {
-      return;
-    }
-
-    // Collect each detail's errors object into single array of error messages.
-    let detailsErrorMessages = Ember.A();
-    details.forEach((detail, i) => {
-      let detailErrors = Ember.get(detail, 'errors');
-
-      for (let detailPropertyName in detailErrors) {
-        let detailPropertyErrorMessages = detailErrors[detailPropertyName];
-        if (detailErrors.hasOwnProperty(detailPropertyName) && Ember.isArray(detailPropertyErrorMessages)) {
-          detailPropertyErrorMessages.forEach((detailPropertyErrorMessage) => {
-            Ember.removeObserver(this, `${detailsName}.@each.${detailPropertyName}`, this, this._onChangeHasManyRelationship);
-
-            if (!Ember.get(detail, 'isDeleted')) {
-              Ember.addObserver(this, `${detailsName}.@each.${detailPropertyName}`, this, this._onChangeHasManyRelationship);
-              detailsErrorMessages.pushObject(detailPropertyErrorMessage);
-            }
-          });
-        }
-      }
-    });
-
-    // Remember array of error messages in aggregator's errors object.
-    Ember.set(errors, detailsName, detailsErrorMessages);
   },
 
   /**
@@ -518,27 +362,27 @@ var Model = DS.Model.extend(EmberValidations, Ember.Evented, {
 
 Model.reopenClass({
   /**
-   * Defined projections for current model type.
-   *
-   * @property projections
-   * @type Ember.Object
-   * @default null
-   * @public
-   * @static
-   */
+    Defined projections for current model type.
+
+    @property projections
+    @type Ember.Object
+    @default null
+    @public
+    @static
+  */
   projections: null,
 
   /**
-   * Defines projection for specified model type.
-   *
-   * @method defineProjection
-   * @param {String} projectionName Projection name, eg 'EmployeeE'.
-   * @param {String} modelName The name of the model type.
-   * @param {Object} attributes Projection attributes.
-   * @return {Object} Created projection.
-   * @public
-   * @static
-   */
+    Defines projection for specified model type.
+
+    @method defineProjection
+    @param {String} projectionName Projection name, eg 'EmployeeE'.
+    @param {String} modelName The name of the model type.
+    @param {Object} attributes Projection attributes.
+    @return {Object} Created projection.
+    @public
+    @static
+  */
   defineProjection: function (projectionName, modelName, attributes) {
     let proj = createProj(modelName, attributes, projectionName);
 
@@ -558,14 +402,14 @@ Model.reopenClass({
   },
 
   /**
-   * Parent model type name.
-   *
-   * @property _parentModelName
-   * @type String
-   * @default null
-   * @private
-   * @static
-   */
+    Parent model type name.
+
+    @property _parentModelName
+    @type String
+    @default null
+    @private
+    @static
+  */
   _parentModelName: null
 });