+++ tests/dummy/app/models/ember-flexberry-dummy-application-user.js
@@ -16,20 +16,6 @@ let Model = Projection.Model.extend(Offline.ModelMixin, {
   vip: DS.attr('boolean'),
   karma: DS.attr('decimal'),
   coordinates: DS.attr('string'),
-
-  // Model validation rules.
-  validations: {
-    name: {
-      presence: {
-        message: 'Name is required'
-      }
-    },
-    eMail: {
-      presence: {
-        message: 'User email is required'
-      }
-    }
-  }
 });
 
 // Edit form projection.