+++ tests/dummy/app/models/ember-flexberry-dummy-test-poly.js
@@ -4,15 +4,6 @@ import { Projection } from 'ember-flexberry-data';
 let Model = Projection.Model.extend({
   selfPole: DS.attr('string'),
   relation: DS.belongsTo('ember-flexberry-dummy-test-poly-base', { inverse: null, async: false, polymorphic: true }),
-
-  // Model validation rules.
-  validations: {
-    relation: {
-      presence: {
-        message: 'Relation is required'
-      }
-    }
-  }
 });
 
 Model.defineProjection('TestPolyEdit', 'ember-flexberry-dummy-test-poly', {