+++ tests/dummy/app/models/ember-flexberry-dummy-vote.js
@@ -12,10 +12,6 @@ var Model = Projection.Model.extend({
 
   // This property is for flexberry-lookup component. No inverse relationship here.
   author: DS.belongsTo('ember-flexberry-dummy-application-user', { inverse: null, async: false }),
-
-  // Model validation rules.
-  validations: {
-  }
 });
 
 // Edit form projection.